+++ frontend/src/api/index.ts
@@ -1,4 +1,4 @@
-import client from './client';
+import client from '../lib/client';
 
 /**
  * API 客户端测试函数
@@ -17,7 +17,7 @@ export const testApiClient = async () => {
 };
 
 // 导出客户端供其他模块使用
-export { default } from './client';
+export { default } from '../lib/client';
 
 // 导出 API 工具函数
 export * from '../lib/api';
